put interface to right location and add test script in the comment section
