+++ next-env.d.ts
@@ -9,3 +9,15 @@ interface EventData {
   Timestamp: Date;
   [Field: string]: string;
 }
+
+// The data of the event which needs to be passed in during the api call.
+interface CalendarEventData {
+  Name: string;
+  Description: string;
+  Organization: string;
+  Location: string;
+  StartDate: string;    // format according to RFC5545, use toISOString() before send the api request
+  EndDate: string;      // format according to RFC5545, use toISOString() before send the api request
+  Timezone: string;     // Formatted as an IANA Time Zone Database name, e.g. "Europe/Zurich"
+  Recurrence?: string[]; // format according to RFC5545
+}
